server/pyproject.toml
[project]
name = "ai-ppt-agent-server"
version = "0.1.0"
description = "AI PPT Agent 后端服务 — LLM 驱动的演示文稿生成引擎"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.18",

    # Data Validation
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",

    # LLM SDKs
    "openai>=1.60.0",
    "google-genai>=1.0.0",

    # Document Parsing
    "python-pptx>=1.0.0",
    "python-docx>=1.1.0",
    "pymupdf>=1.25.0",

    # SVG Processing
    "lxml>=5.3.0",

    # Search Integration
    "httpx>=0.28.0",

    # Async Task Queue
    "celery[redis]>=5.4.0",

    # Caching & Session
    "redis>=5.2.0",

    # JSON Repair (LLM output)
    "json-repair>=0.30.0",

    # Utilities
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
